+++ BackEnd/IdeaManager-API/src/routes/idea.routes.ts
@@ -1,7 +1,21 @@
 import {Router,Request,Response} from 'express'
+import {} from "../services/ideaServices"
 
 
 const idearouter = Router()
+
+idearouter.get('/',(request,response)=>{
+    try{
+        
+    }catch(error){
+        response.status(400).json(error.message)
+    }
+})
+
+
+idearouter.get('/:id',(request,response) =>{
+    let {id} = request.params
+})