+++ BackEnd/IdeaManager-API/src/services/ideaServices/index.ts
@@ -1,0 +1,11 @@
+import CreateIdeaService from './CreateIdeaService'
+import UpdateIdeaByNameService from './UpdateIdeaByNameService'
+
+class IdeaServices{
+    // let CreateIdeaService: CreateIdeaService;
+    // let UpdateIdeaByNameService: UpdateIdeaByNameService
+
+    // createor
+}
+
+export default IdeaServices