+++ Changelog.md
@@ -1,3 +1,29 @@
+# [0.0.2] - 30/08/2020
+## Inclusion of Json in all routes
+## Creation of Idea Services
+## Creation of Idea Creation Service
+## Creation of Idea Update Service by Name
+## Route to return all Ideas
+## Route to return Idea by ID
+
+**Changed Files**
+
+        modified:   BackEnd/IdeaManager-API/package.json
+        new file:   BackEnd/IdeaManager-API/src/models/Idea.ts
+        new file:   BackEnd/IdeaManager-API/src/models/Tech.ts
+        new file:   BackEnd/IdeaManager-API/src/repositories/IdeaRepository.ts
+        modified:   BackEnd/IdeaManager-API/src/routes/idea.routes.ts
+        modified:   BackEnd/IdeaManager-API/src/routes/index.ts
+        modified:   BackEnd/IdeaManager-API/src/server.ts
+        new file:   BackEnd/IdeaManager-API/src/services/ideaServices/CreateIdeaService.ts
+        new file:   BackEnd/IdeaManager-API/src/services/ideaServices/UpdateIdeaByNameService.ts
+        new file:   BackEnd/IdeaManager-API/src/services/ideaServices/index.ts
+        new file:   BackEnd/IdeaManager-API/yarn-error.log
+        modified:   BackEnd/IdeaManager-API/yarn.lock
+        modified:   Changelog.md
+
+---
+
 # [0.0.1] - 30/08/2020
 ## Initial Commmit 
 ## Folder Structure Creation 
@@ -8,3 +34,13 @@
 
 
 **Changed Files**
+
+        new file:   .gitignore
+        new file:   BackEnd/IdeaManager-API/package.json
+        new file:   BackEnd/IdeaManager-API/src/routes/idea.routes.ts
+        new file:   BackEnd/IdeaManager-API/src/routes/index.ts
+        new file:   BackEnd/IdeaManager-API/src/server.ts
+        new file:   BackEnd/IdeaManager-API/src/services/CreateMessageLogRoutesService.ts
+        new file:   BackEnd/IdeaManager-API/tsconfig.json
+        new file:   BackEnd/IdeaManager-API/yarn.lock
+        new file:   Changelog.md